feat: reorganize event classes into subdirectories for better structure
