+++ src/Events/Folders/FolderDeleted.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Mindtwo\LaravelClickUpApi\Events;
+namespace Mindtwo\LaravelClickUpApi\Events\Folders;
+
+use Mindtwo\LaravelClickUpApi\Events\ClickUpEvent;
 
 class FolderDeleted extends ClickUpEvent
 {
