--- src/Events/GoalUpdated.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Mindtwo\LaravelClickUpApi\Events;
-
-class GoalUpdated extends ClickUpEvent {}
