+++ src/Events/Goals/GoalCreated.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mindtwo\LaravelClickUpApi\Events\Goals;
+
+use Mindtwo\LaravelClickUpApi\Events\ClickUpEvent;
+
+class GoalCreated extends ClickUpEvent {}
