+++ src/Events/Goals/GoalDeleted.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mindtwo\LaravelClickUpApi\Events\Goals;
+
+use Mindtwo\LaravelClickUpApi\Events\ClickUpEvent;
+
+class GoalDeleted extends ClickUpEvent {}
