+++ src/Events/Goals/GoalUpdated.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mindtwo\LaravelClickUpApi\Events\Goals;
+
+use Mindtwo\LaravelClickUpApi\Events\ClickUpEvent;
+
+class GoalUpdated extends ClickUpEvent {}
