--- src/Events/KeyResultCreated.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Mindtwo\LaravelClickUpApi\Events;
-
-class KeyResultCreated extends ClickUpEvent {}
