+++ src/Events/KeyResults/KeyResultUpdated.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mindtwo\LaravelClickUpApi\Events\KeyResults;
+
+use Mindtwo\LaravelClickUpApi\Events\ClickUpEvent;
+
+class KeyResultUpdated extends ClickUpEvent {}
