+++ src/Events/Tasks/TaskPriorityUpdated.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Mindtwo\LaravelClickUpApi\Events;
+namespace Mindtwo\LaravelClickUpApi\Events\Tasks;
 
+use Mindtwo\LaravelClickUpApi\Events\ClickUpEvent;
 use Mindtwo\LaravelClickUpApi\Traits\HandlesTask;
 
 class TaskPriorityUpdated extends ClickUpEvent
