+++ src/Http/Controllers/WebhookController.php
@@ -11,34 +11,34 @@ use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
 use Mindtwo\LaravelClickUpApi\Enums\EventSource;
-use Mindtwo\LaravelClickUpApi\Events\FolderCreated;
-use Mindtwo\LaravelClickUpApi\Events\FolderDeleted;
-use Mindtwo\LaravelClickUpApi\Events\FolderUpdated;
-use Mindtwo\LaravelClickUpApi\Events\GoalCreated;
-use Mindtwo\LaravelClickUpApi\Events\GoalDeleted;
-use Mindtwo\LaravelClickUpApi\Events\GoalUpdated;
-use Mindtwo\LaravelClickUpApi\Events\KeyResultCreated;
-use Mindtwo\LaravelClickUpApi\Events\KeyResultDeleted;
-use Mindtwo\LaravelClickUpApi\Events\KeyResultUpdated;
-use Mindtwo\LaravelClickUpApi\Events\ListCreated;
-use Mindtwo\LaravelClickUpApi\Events\ListDeleted;
-use Mindtwo\LaravelClickUpApi\Events\ListUpdated;
-use Mindtwo\LaravelClickUpApi\Events\SpaceCreated;
-use Mindtwo\LaravelClickUpApi\Events\SpaceDeleted;
-use Mindtwo\LaravelClickUpApi\Events\SpaceUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskAssigneeUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskCommentPosted;
-use Mindtwo\LaravelClickUpApi\Events\TaskCommentUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskCreated;
-use Mindtwo\LaravelClickUpApi\Events\TaskDeleted;
-use Mindtwo\LaravelClickUpApi\Events\TaskDueDateUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskMoved;
-use Mindtwo\LaravelClickUpApi\Events\TaskPriorityUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskStatusUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskTagUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskTimeEstimateUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskTimeTrackedUpdated;
-use Mindtwo\LaravelClickUpApi\Events\TaskUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Folders\FolderCreated;
+use Mindtwo\LaravelClickUpApi\Events\Folders\FolderDeleted;
+use Mindtwo\LaravelClickUpApi\Events\Folders\FolderUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Goals\GoalCreated;
+use Mindtwo\LaravelClickUpApi\Events\Goals\GoalDeleted;
+use Mindtwo\LaravelClickUpApi\Events\Goals\GoalUpdated;
+use Mindtwo\LaravelClickUpApi\Events\KeyResults\KeyResultCreated;
+use Mindtwo\LaravelClickUpApi\Events\KeyResults\KeyResultDeleted;
+use Mindtwo\LaravelClickUpApi\Events\KeyResults\KeyResultUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Lists\ListCreated;
+use Mindtwo\LaravelClickUpApi\Events\Lists\ListDeleted;
+use Mindtwo\LaravelClickUpApi\Events\Lists\ListUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Spaces\SpaceCreated;
+use Mindtwo\LaravelClickUpApi\Events\Spaces\SpaceDeleted;
+use Mindtwo\LaravelClickUpApi\Events\Spaces\SpaceUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskAssigneeUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskCommentPosted;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskCommentUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskCreated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskDeleted;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskDueDateUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskMoved;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskPriorityUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskStatusUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskTagUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskTimeEstimateUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskTimeTrackedUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskUpdated;
 use Mindtwo\LaravelClickUpApi\Models\ClickUpWebhook;
 use Mindtwo\LaravelClickUpApi\Models\ClickUpWebhookDelivery;
