+++ src/Jobs/ClickUpApiCallJob.php
@@ -15,9 +15,9 @@ use Illuminate\Queue\SerializesModels;
 use Mindtwo\LaravelClickUpApi\ClickUpClient;
 use Mindtwo\LaravelClickUpApi\Enums\EventSource;
 use Mindtwo\LaravelClickUpApi\Events\ClickUpApiCallCompleted;
-use Mindtwo\LaravelClickUpApi\Events\TaskCreated;
-use Mindtwo\LaravelClickUpApi\Events\TaskDeleted;
-use Mindtwo\LaravelClickUpApi\Events\TaskUpdated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskCreated;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskDeleted;
+use Mindtwo\LaravelClickUpApi\Events\Tasks\TaskUpdated;
 use Mindtwo\LaravelClickUpApi\Http\LazyResponseProxy;
 use RuntimeException;
 use Symfony\Component\HttpFoundation\Request;
